Refactor(Player):ComboSystemFeatureComplet
Prototype of combo system, the user can create attacks and combos for the player.
The system is unstable and can crash if the player forgets to correctly set attacks in the ComboTree, and soft-lock the player if he do not add Notify in the animation.
Add Animation Pack to test the system (the pack is only a placeholder).
Work in Progress.